rewrite of "request" events
needed to simplify fix of #280.

RestApiUser uses locust.contrib.fasthttp.FastHttpSession for synchronous requests as well.

RequestLogger and RequestHandler is not inherited, but always added for users inheriting GrizzlyUser, and added as event listeners on the dedicated event hook.

A new GrizzlyEventHook, so let through all exceptions.

ResponseContextManager is only used with in RestApiUser._request, all events are fired from GrizzlyUser.request (final), so it's done the same for all users.
