+++ grizzly/users/servicebus.py
@@ -83,7 +83,7 @@ from grizzly.utils import is_template
 from grizzly_extras.arguments import get_unsupported_arguments, parse_arguments
 from grizzly_extras.async_message import AsyncMessageContext, AsyncMessageRequest, AsyncMessageResponse, async_message_request
 
-from .base import GrizzlyUser, ResponseHandler, grizzlycontext
+from .base import GrizzlyUser, grizzlycontext
 
 MAX_LENGTH = 65
 
@@ -93,7 +93,7 @@ MAX_LENGTH = 65
         'wait': None,
     },
 })
-class ServiceBusUser(ResponseHandler, GrizzlyUser):
+class ServiceBusUser(GrizzlyUser):
     __dependencies__: ClassVar[Set[str]] = {'async-messaged'}
 
     am_context: AsyncMessageContext
