+++ tests/unit/test_grizzly/events/__init__.py
@@ -1,0 +1,1 @@
+"""Commence initialization of module."""
